+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Bug Fixes
 - Selection area on taking screenshot or switching between screeenshots is not removed!
+- Selection with Ctrl+A not worked after clicking on toolbar or bottom area.
 
 # 2025.01.11