+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ﻿# 2025.01.23
-(unpublished)
 
 ## Changes
 - V-cut, H-cut icons were changed.
@@ -8,6 +7,7 @@
 - Selection area on taking screenshot or switching between screeenshots is not removed!
 - Selection with Ctrl+A not worked after clicking on toolbar or bottom area.
 - During taking screenshot previous active project was not saved and changes were lost.
+- Cut out areas of image were not saved.
 
 # 2025.01.11